Replace Travis CI with GitHub workflow.
Travis no longer supports free builds for open source project backed by
companies.  We therefore switch to using GitHub actions instead.  Note
that most of the tests still run in GitLab and should probably be ported
to GitHub actions in the future.  For now we simply replicate the Travis
workflow.